Script Engine Preprocessor bug fixed
