+++ FreeStrange/ScriptEngine/VM.py
@@ -3,6 +3,7 @@ Virtual Machine
 """
 import sys
 from collections import deque
+from . import keyword_list, wait_list
 
 
 class Stack(deque):
@@ -188,3 +189,23 @@ class VirtualMachine:
 
     def use(self):
         pass
+
+
+class Preprocessor:
+    def __init__(self, code):
+        self.stack = []
+        self.out = []
+        wait = []
+        for i in code:
+            if len(wait) != 0:
+                self.out += self.stack
+                self.out.append(i)
+                self.out += wait
+                wait.pop()
+                self.stack = []
+            elif (i in keyword_list) and (i in wait_list):
+                wait.append(i)
+            elif (i in keyword_list) and (i not in wait_list):
+                self.out.append(i)
+            else:
+                self.stack.append(i)
